:sparkles: feat: 서버 - 클라이언트 멀티플레이 기초 구축
- 두명의 클라이언트가 들어오면 게임이 시작, 창 띄움
- 클라이언트 키보트 이벤트에 따라서 클라이언트 -> 서버 통신 구축 필요
