+++ srcs/Server/includes/CircleObject.hpp
