+++ srcs/Server/includes/Socket.hpp
@@ -1,0 +1,38 @@
+#ifndef SOCKET_HPP
+#define SOCKET_HPP
+
+#define MAX_CLIENTS 1024
+#define BUFFER_SIZE 1024
+
+#include <unistd.h>
+#include <cstdlib>
+#include <iostream>
+#include <cstring>
+#include <sys/socket.h>
+
+
+enum e_error
+{
+    WRONG_ARG,
+    FATAL_ERR,
+};
+
+enum e_msg
+{
+    ARRIVE,
+    LEFT,
+    MSG,
+};
+
+struct s_Client
+{
+    int id, fd;
+    char *buff;
+};
+
+void error_msg(enum e_error flag);
+void send_all(int my_id, enum e_msg flag, char *msg, s_Client *clients);
+int extract_message(char **buf, char **msg);
+char *str_join(char *buf, char *add);
+
+#endif // SOCKET_HPP
